Added Program
Added a program where the user is checked if they are eligible for a job. The requirements is that they must be a US Citizen, have a bachelors degree OR 2 years of experience. With the combination of logical operators and if statements, the code will determine if one is eligible